requirements.txt
requests
pyjwt
numpy
